+++ sway-core/src/semantic_analysis/ast_node/mod.rs
@@ -7,7 +7,12 @@ pub(crate) use expression::*;
 pub(crate) use modes::*;
 
 use crate::{
-    language::{parsed::*, ty},
+    engine_threading::SpannedWithEngines,
+    language::{
+        parsed::*,
+        ty::{self, TyDecl},
+    },
+    namespace::TraitMap,
     semantic_analysis::*,
     type_system::*,
     Engines, Ident,
@@ -52,6 +57,7 @@ impl ty::TyAstNode {
             content: match node.content.clone() {
                 AstNodeContent::UseStatement(stmt) => {
                     handle_use_statement(&mut ctx, engines, &stmt, handler);
+                    handle_item_trait_imports(&mut ctx, engines, handler)?;
                     ty::TyAstNodeContent::SideEffect(ty::TySideEffect {
                         side_effect: ty::TySideEffectVariant::UseStatement(ty::TyUseStatement {
                             alias: stmt.alias,
@@ -128,6 +134,54 @@ impl ty::TyAstNode {
 
         Ok(node)
     }
+}
+
+fn handle_item_trait_imports(
+    ctx: &mut TypeCheckContext<'_>,
+    engines: &Engines,
+    handler: &Handler,
+) -> Result<(), ErrorEmitted> {
+    let mut impls_to_insert = TraitMap::default();
+
+    let root_mod = &ctx.namespace().root().module;
+    let dst_mod = ctx.namespace.module(engines);
+
+    for (_, (_, src, decl)) in dst_mod.current_items().use_item_synonyms.iter() {
+        let src_mod = root_mod.lookup_submodule(handler, engines, src)?;
+
+        //  if this is an enum or struct or function, import its implementations
+        if let Ok(type_id) = decl.return_type(&Handler::default(), engines) {
+            impls_to_insert.extend(
+                src_mod
+                    .current_items()
+                    .implemented_traits
+                    .filter_by_type_item_import(type_id, engines),
+                engines,
+            );
+        }
+        // if this is a trait, import its implementations
+        let decl_span = decl.span(engines);
+        if matches!(decl, TyDecl::TraitDecl(_)) {
+            // TODO: we only import local impls from the source namespace
+            // this is okay for now but we'll need to device some mechanism to collect all
+            // available trait impls
+            impls_to_insert.extend(
+                src_mod
+                    .current_items()
+                    .implemented_traits
+                    .filter_by_trait_decl_span(decl_span),
+                engines,
+            );
+        }
+    }
+
+    let dst_mod = ctx.namespace_mut().module_mut(engines);
+    dst_mod
+        .current_items_mut()
+        .implemented_traits
+        .extend(impls_to_insert, engines);
+
+    Ok(())
 }
 
 fn handle_use_statement(
