+++ sway-core/src/semantic_analysis/namespace/root.rs
@@ -1,8 +1,6 @@
 use std::fmt;
 
-use super::{
-    module::Module, namespace::Namespace, trait_map::TraitMap, Ident, ResolvedTraitImplItem,
-};
+use super::{module::Module, namespace::Namespace, Ident, ResolvedTraitImplItem};
 use crate::{
     decl_engine::DeclRef,
     engine_threading::*,
@@ -64,35 +62,6 @@ impl ResolvedDeclaration {
         match self {
             ResolvedDeclaration::Parsed(decl) => decl.visibility(engines.pe()),
             ResolvedDeclaration::Typed(decl) => decl.visibility(engines.de()),
-        }
-    }
-
-    fn span(&self, engines: &Engines) -> sway_types::Span {
-        match self {
-            ResolvedDeclaration::Parsed(decl) => decl.span(engines),
-            ResolvedDeclaration::Typed(decl) => decl.span(engines),
-        }
-    }
-
-    pub(crate) fn return_type(
-        &self,
-        handler: &Handler,
-        engines: &Engines,
-    ) -> Result<TypeId, ErrorEmitted> {
-        match self {
-            ResolvedDeclaration::Parsed(_decl) => unreachable!(),
-            ResolvedDeclaration::Typed(decl) => decl.return_type(handler, engines),
-        }
-    }
-
-    fn is_trait(&self) -> bool {
-        match self {
-            ResolvedDeclaration::Parsed(decl) => {
-                matches!(decl, Declaration::TraitDeclaration(_))
-            }
-            ResolvedDeclaration::Typed(decl) => {
-                matches!(decl, TyDecl::TraitDecl(_))
-            }
         }
     }
 }
@@ -188,7 +157,6 @@ impl Root {
         self.check_module_privacy(handler, engines, src)?;
 
         let src_mod = self.module.lookup_submodule(handler, engines, src)?;
-        let mut impls_to_insert = TraitMap::default();
         match src_mod.current_items().symbols.get(item).cloned() {
             Some(decl) => {
                 if !decl.visibility(engines).is_public() && !is_ancestor(src, dst) {
@@ -198,29 +166,6 @@ impl Root {
                     });
                 }
 
-                //  if this is an enum or struct or function, import its implementations
-                if let Ok(type_id) = decl.return_type(&Handler::default(), engines) {
-                    impls_to_insert.extend(
-                        src_mod
-                            .current_items()
-                            .implemented_traits
-                            .filter_by_type_item_import(type_id, engines),
-                        engines,
-                    );
-                }
-                // if this is a trait, import its implementations
-                let decl_span = decl.span(engines);
-                if decl.is_trait() {
-                    // TODO: we only import local impls from the source namespace
-                    // this is okay for now but we'll need to device some mechanism to collect all available trait impls
-                    impls_to_insert.extend(
-                        src_mod
-                            .current_items()
-                            .implemented_traits
-                            .filter_by_trait_decl_span(decl_span),
-                        engines,
-                    );
-                }
                 // no matter what, import it this way though.
                 let dst_mod = self.module.lookup_submodule_mut(handler, engines, dst)?;
                 let check_name_clash = |name| {
@@ -253,12 +198,6 @@ impl Root {
                 }));
             }
         };
-
-        let dst_mod = self.module.lookup_submodule_mut(handler, engines, dst)?;
-        dst_mod
-            .current_items_mut()
-            .implemented_traits
-            .extend(impls_to_insert, engines);
 
         Ok(())
     }
